@@ -1,6 +1,6 @@
 {
   "name": "generator-angular-fullstack",
-  "version": "3.7.6",
+  "version": "4.0.0-rc.0",
   "description": "Yeoman generator for creating MEAN stack applications, using MongoDB, Express, AngularJS, and Node",
   "keywords": [
     "yeoman-generator",
@@ -40,11 +40,12 @@
     "babel-plugin-transform-flow-strip-types": "^6.7.0",
     "bluebird": "^3.3.5",
     "chalk": "^1.1.0",
-    "generator-ng-component": "~0.3.1",
+    "generator-ng-component": "~1.0.1",
     "glob": "^7.0.3",
     "gulp-babel": "^6.1.2",
     "gulp-beautify": "^2.0.0",
     "gulp-filter": "^4.0.0",
+    "gulp-tap": "^0.1.3",
     "insight": "~0.8.1",
     "lodash": "^4.13.1",
     "semver": "^5.1.0",
@@ -73,7 +74,6 @@
     "gulp-plumber": "^1.1.0",
     "gulp-util": "^3.0.7",
     "jit-grunt": "~0.10.0",
-    "jscs": "^3.0.3",
     "lazypipe": "^1.0.1",
     "merge-stream": "^1.0.0",
     "minimatch": "^3.0.2",
@@ -87,8 +87,8 @@
     "yeoman-test": "~1.4.0"
   },
   "engines": {
-    "node": "^5.10.1",
-    "npm": "^3.8.3"
+    "node": "^6.2.2",
+    "npm": "^3.9.5"
   },
   "license": "BSD-2-Clause"
 }
