@@ -1,13 +1,13 @@
 'use strict';
 
-import crypto from 'crypto';<% if (filters.oauth) { %>
+import crypto from 'crypto';<% if(filters.oauth) { %>
 var authTypes = ['github', 'twitter', 'facebook', 'google'];<% } %>
 
 var validatePresenceOf = function(value) {
   return value && value.length;
 };
 
-module.exports = function(sequelize, DataTypes) {
+export default function(sequelize, DataTypes) {
   var User = sequelize.define('User', {
 
     _id: {
@@ -37,9 +37,9 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     provider: DataTypes.STRING,
-    salt: DataTypes.STRING<% if (filters.oauth) { %>,<% if (filters.facebookAuth) { %>
-    facebook: DataTypes.JSON,<% } %><% if (filters.twitterAuth) { %>
-    twitter: DataTypes.JSON,<% } %><% if (filters.googleAuth) { %>
+    salt: DataTypes.STRING<% if(filters.oauth) { %>,<% if(filters.facebookAuth) { %>
+    facebook: DataTypes.JSON,<% } %><% if(filters.twitterAuth) { %>
+    twitter: DataTypes.JSON,<% } %><% if(filters.googleAuth) { %>
     google: DataTypes.JSON,<% } %>
     github: DataTypes.JSON<% } %>
 
@@ -52,16 +52,16 @@ module.exports = function(sequelize, DataTypes) {
       // Public profile information
       profile: function() {
         return {
-          'name': this.name,
-          'role': this.role
+          name: this.name,
+          role: this.role
         };
       },
 
       // Non-sensitive info we'll be putting in the token
       token: function() {
         return {
-          '_id': this._id,
-          'role': this.role
+          _id: this._id,
+          role: this.role
         };
       }
     },
@@ -70,25 +70,25 @@ module.exports = function(sequelize, DataTypes) {
      * Pre-save hooks
      */
     hooks: {
-      beforeBulkCreate: function(users, fields, fn) {
+      beforeBulkCreate(users, fields, fn) {
         var totalUpdated = 0;
-        users.forEach(function(user) {
-          user.updatePassword(function(err) {
-            if (err) {
+        users.forEach(user => {
+          user.updatePassword(err => {
+            if(err) {
               return fn(err);
             }
             totalUpdated += 1;
-            if (totalUpdated === users.length) {
+            if(totalUpdated === users.length) {
               return fn();
             }
           });
         });
       },
-      beforeCreate: function(user, fields, fn) {
+      beforeCreate(user, fields, fn) {
         user.updatePassword(fn);
       },
-      beforeUpdate: function(user, fields, fn) {
-        if (user.changed('password')) {
+      beforeUpdate(user, fields, fn) {
+        if(user.changed('password')) {
           return user.updatePassword(fn);
         }
         fn();
@@ -107,18 +107,18 @@ module.exports = function(sequelize, DataTypes) {
        * @return {Boolean}
        * @api public
        */
-      authenticate: function(password, callback) {
-        if (!callback) {
+      authenticate(password, callback) {
+        if(!callback) {
           return this.password === this.encryptPassword(password);
         }
 
         var _this = this;
         this.encryptPassword(password, function(err, pwdGen) {
-          if (err) {
+          if(err) {
             callback(err);
           }
 
-          if (_this.password === pwdGen) {
+          if(_this.password === pwdGen) {
             callback(null, true);
           }
           else {
@@ -130,32 +130,29 @@ module.exports = function(sequelize, DataTypes) {
       /**
        * Make salt
        *
-       * @param {Number} byteSize Optional salt byte size, default to 16
+       * @param {Number} [byteSize] - Optional salt byte size, default to 16
        * @param {Function} callback
        * @return {String}
        * @api public
        */
-      makeSalt: function(byteSize, callback) {
+      makeSalt(byteSize, callback) {
         var defaultByteSize = 16;
 
-        if (typeof arguments[0] === 'function') {
+        if(typeof arguments[0] === 'function') {
           callback = arguments[0];
           byteSize = defaultByteSize;
-        }
-        else if (typeof arguments[1] === 'function') {
+        } else if(typeof arguments[1] === 'function') {
           callback = arguments[1];
+        } else {
+          throw new Error('Missing Callback');
         }
 
-        if (!byteSize) {
+        if(!byteSize) {
           byteSize = defaultByteSize;
-        }
-
-        if (!callback) {
-          return crypto.randomBytes(byteSize).toString('base64');
         }
 
         return crypto.randomBytes(byteSize, function(err, salt) {
-          if (err) {
+          if(err) {
             callback(err);
           }
           return callback(null, salt.toString('base64'));
@@ -170,26 +167,23 @@ module.exports = function(sequelize, DataTypes) {
        * @return {String}
        * @api public
        */
-      encryptPassword: function(password, callback) {
-        if (!password || !this.salt) {
-          if (!callback) {
-            return null;
-          }
-          return callback(null);
+      encryptPassword(password, callback) {
+        if(!password || !this.salt) {
+          return callback ? callback(null) : null;
         }
 
         var defaultIterations = 10000;
         var defaultKeyLength = 64;
         var salt = new Buffer(this.salt, 'base64');
 
-        if (!callback) {
+        if(!callback) {
           return crypto.pbkdf2Sync(password, salt, defaultIterations, defaultKeyLength)
                        .toString('base64');
         }
 
         return crypto.pbkdf2(password, salt, defaultIterations, defaultKeyLength,
           function(err, key) {
-            if (err) {
+            if(err) {
               callback(err);
             }
             return callback(null, key.toString('base64'));
@@ -203,31 +197,28 @@ module.exports = function(sequelize, DataTypes) {
        * @return {String}
        * @api public
        */
-      updatePassword: function(fn) {
+      updatePassword(fn) {
         // Handle new/update passwords
-        if (this.password) {
-          if (!validatePresenceOf(this.password)<% if (filters.oauth) { %> && authTypes.indexOf(this.provider) === -1<% } %>) {
-            fn(new Error('Invalid password'));
-          }
+        if(!this.password) return fn(null);
 
-          // Make salt with a callback
-          var _this = this;
-          this.makeSalt(function(saltErr, salt) {
-            if (saltErr) {
-              fn(saltErr);
-            }
-            _this.salt = salt;
-            _this.encryptPassword(_this.password, function(encryptErr, hashedPassword) {
-              if (encryptErr) {
-                fn(encryptErr);
-              }
-              _this.password = hashedPassword;
-              fn(null);
-            });
-          });
-        } else {
-          fn(null);
+        if(!validatePresenceOf(this.password)<% if(filters.oauth) { %> && authTypes.indexOf(this.provider) === -1<% } %>) {
+          fn(new Error('Invalid password'));
         }
+
+        // Make salt with a callback
+        this.makeSalt((saltErr, salt) => {
+          if(saltErr) {
+            return fn(saltErr);
+          }
+          this.salt = salt;
+          this.encryptPassword(this.password, (encryptErr, hashedPassword) => {
+            if(encryptErr) {
+              fn(encryptErr);
+            }
+            this.password = hashedPassword;
+            fn(null);
+          });
+        });
       }
     }
   });
