@@ -11,6 +11,7 @@ import insight from '../insight-init';
 import {exec} from 'child_process';
 import babelStream from 'gulp-babel';
 import beaufityStream from 'gulp-beautify';
+import tap from 'gulp-tap';
 import filter from 'gulp-filter';
 import semver from 'semver';
 
@@ -22,6 +23,12 @@ export class Generator extends Base {
 
     this.option('skip-install', {
       desc: 'Do not install dependencies',
+      type: Boolean,
+      defaults: false
+    });
+
+    this.option('skip-config', {
+      desc: 'Always use existing .yo-rc.json',
       type: Boolean,
       defaults: false
     });
@@ -93,34 +100,38 @@ export class Generator extends Base {
       checkForConfig: function() {
         var existingFilters = this.config.get('filters');
 
-        if(existingFilters) {
-          return this.prompt([{
-            type: 'confirm',
-            name: 'skipConfig',
-            message: 'Existing .yo-rc configuration found, would you like to use it?',
-            default: true,
-          }]).then(answers => {
-            this.skipConfig = answers.skipConfig;
+        if(!existingFilters) return;
 
-            if(this.skipConfig) {
-              insight.track('skipConfig', 'true');
-              this.filters = existingFilters;
+        let promise = this.options['skip-config']
+          ? Promise.resolve({skipConfig: true})
+          : this.prompt([{
+              type: 'confirm',
+              name: 'skipConfig',
+              message: 'Existing .yo-rc configuration found, would you like to use it?',
+              default: true,
+            }]);
 
-              this.scriptExt = this.filters.ts ? 'ts' : 'js';
-              this.templateExt = this.filters.jade ? 'jade' : 'html';
-              this.styleExt = this.filters.sass ? 'scss' :
-                this.filters.less ? 'less' :
-                this.filters.stylus ? 'styl' :
-                'css';
-            } else {
-              insight.track('skipConfig', 'false');
-              this.filters = {};
-              this.forceConfig = true;
-              this.config.set('filters', this.filters);
-              this.config.forceSave();
-            }
-          });
-        }
+        return promise.then(answers => {
+          this.skipConfig = answers.skipConfig;
+
+          if(this.skipConfig) {
+            insight.track('skipConfig', 'true');
+            this.filters = existingFilters;
+
+            this.scriptExt = this.filters.ts ? 'ts' : 'js';
+            this.templateExt = this.filters.jade ? 'jade' : 'html';
+            this.styleExt = this.filters.sass ? 'scss' :
+              this.filters.less ? 'less' :
+              this.filters.stylus ? 'styl' :
+              'css';
+          } else {
+            insight.track('skipConfig', 'false');
+            this.filters = {};
+            this.forceConfig = true;
+            this.config.set('filters', this.filters);
+            this.config.forceSave();
+          }
+        });
       },
       assignPorts() {
         this.devPort = this.options['dev-port'];
@@ -149,12 +160,12 @@ export class Generator extends Base {
               }[val];
             }
           }, {
-          // TODO: enable once Babel setup supports Flow
-          //   type: 'confirm',
-          //   name: 'flow',
-          //   message: 'Would you like to use Flow types with Babel?',
-          //   when: answers => answers.transpiler === 'babel'
-          // }, {
+            type: 'confirm',
+            name: 'flow',
+            default: false,
+            message: 'Would you like to use Flow types with Babel?',
+            when: answers => answers.transpiler === 'babel'
+          }, {
             type: 'list',
             name: 'markup',
             message: 'What would you like to write markup with?',
@@ -318,16 +329,6 @@ export class Generator extends Base {
 
         return this.prompt([{
           type: 'list',
-          name: 'buildtool',
-          message: 'Would you like to use Gulp or Grunt?',
-          choices: ['Gulp', 'Grunt (deprecated)'],
-          default: 0,
-          filter: val => ({
-            'Gulp': 'gulp',
-            'Grunt (deprecated)': 'grunt'
-          })[val]
-        }, {
-          type: 'list',
           name: 'testing',
           message: 'What would you like to write tests with?',
           choices: ['Jasmine', 'Mocha + Chai + Sinon'],
@@ -346,9 +347,6 @@ export class Generator extends Base {
           filter: val => val.toLowerCase(),
           when: answers => answers.testing === 'mocha'
         }]).then(answers => {
-          this.filters[answers.buildtool] = true;
-          insight.track('buildtool', answers.buildtool);
-
           this.filters[answers.testing] = true;
           insight.track('testing', answers.testing);
           if(answers.testing === 'mocha') {
@@ -417,6 +415,7 @@ export class Generator extends Base {
         if(this.filters.less) extensions.push('less');
 
         filters.push('es6'); // Generate ES6 syntax code
+        filters.push('webpack');  // Generate ES6 Module imports/exports
 
         this.composeWith('ng-component', {
           options: {
@@ -430,7 +429,7 @@ export class Generator extends Base {
             'basePath': 'client',
             'forceConfig': this.forceConfig
           }
-        }, { local: require.resolve('generator-ng-component/app/index.js') });
+        }, { local: require.resolve('generator-ng-component/generators/app/index.js') });
       },
       ngModules: function() {
         var angModules = [
@@ -475,24 +474,32 @@ export class Generator extends Base {
            ]);
          */
 
+        const flow = this.filters.flow;
+
         let babelPlugins = [
           'babel-plugin-syntax-flow',
           'babel-plugin-syntax-class-properties'
         ];
 
-        // TODO: enable once Babel setup supports Flow
-        // if(this.filters.babel && !this.filters.flow) {
+        if(this.filters.babel && !flow) {
           babelPlugins.push('babel-plugin-transform-flow-strip-types');
-        // }
+        }
 
-        const jsFilter = filter(['client/**/*.js'], {restore: true});
+        let jsFilter = filter(['client/**/*.js'], {restore: true});
         this.registerTransformStream([
           jsFilter,
           babelStream({
             plugins: babelPlugins.map(require.resolve),
             /* Babel get's confused about these if you're using an `npm link`ed
                 generator-angular-fullstack, thus the `require.resolve` */
-            // retainLines: true,
+            shouldPrintComment(commentContents) {
+              if(flow) {
+                return true;
+              } else {
+                // strip `// @flow` comments if not using flow
+                return !(/@flow/.test(commentContents));
+              }
+            },
             babelrc: false  // don't grab the generator's `.babelrc`
           }),
           beaufityStream({
@@ -519,6 +526,43 @@ export class Generator extends Base {
           jsFilter.restore
         ]);
 
+        /**
+         * TypeScript doesn't play nicely with things that don't have a default export
+         */
+        if(this.filters.ts) {
+          const modulesToFix = [
+            ['angular', 'angular'],
+            ['ngCookies', 'angular-cookies'],
+            ['ngResource', 'angular-resource'],
+            ['ngSanitize', 'angular-sanitize'],
+            ['uiRouter', 'angular-ui-router'],
+            ['ngRoute', 'angular-route'],
+            ['uiBootstrap', 'angular-ui-bootstrap'],
+            ['ngMessages', 'angular-messages'],
+            ['io', 'socket.io-client']
+          ];
+          function replacer(contents) {
+            modulesToFix.forEach(([moduleName, importName]) => {
+              contents = contents.replace(
+                `import ${moduleName} from '${importName}'`,
+                `const ${moduleName} = require('${importName}')`
+              );
+            });
+            return contents;
+          }
+
+          let tsFilter = filter(['client/**/*.ts'], {restore: true});
+          this.registerTransformStream([
+            tsFilter,
+            tap(function(file, t) {
+              var contents = file.contents.toString();
+              contents = replacer(contents);
+              file.contents = new Buffer(contents);
+            }),
+            tsFilter.restore
+          ]);
+        }
+
         let self = this;
         this.sourceRoot(path.join(__dirname, '../../templates/app'));
         this.processDirectory('.', '.');
@@ -541,14 +585,10 @@ export class Generator extends Base {
     };
   }
 
-  get install() {
-    return {
-      installDeps: function() {
-        this.installDependencies({
-          skipInstall: this.options['skip-install']
-        });
-      }
-    };
+  install() {
+    if(!this.options['skip-install']) {
+      this.spawnCommand('npm', ['install']);
+    }
   }
 
   get end() {
