@@ -6,9 +6,6 @@ import Promise from 'bluebird';
 import helpers from 'yeoman-test';
 import assert from 'yeoman-assert';
 import minimatch from 'minimatch';
-import Checker from 'jscs';
-const jscs = new Checker();
-jscs.registerDefaultRules();
 import * as getExpectedFiles from './get-expected-files';
 import {
   copyAsync,
@@ -21,7 +18,6 @@ import {
 const TEST_DIR = __dirname;
 
 const defaultOptions = {
-  buildtool: 'grunt',
   script: 'js',
   transpiler: 'babel',
   markup: 'html',
@@ -53,12 +49,8 @@ function runEndpointGen(name, opt={}) {
         dir = _dir;
 
         // symlink our dependency directories
-        return Promise.all([
-          fs.mkdirAsync(dir + '/client').then(() => {
-            return fs.symlinkAsync(__dirname + '/fixtures/bower_components', dir + '/client/bower_components');
-          }),
-          fs.symlinkAsync(__dirname + '/fixtures/node_modules', dir + '/node_modules')
-        ]).then(done);
+        return fs.symlinkAsync(__dirname + '/fixtures/node_modules', dir + '/node_modules')
+          .then(done);
       })
       .withOptions(options)
       .withArguments([name])
@@ -75,7 +67,7 @@ function runEndpointGen(name, opt={}) {
   });
 }
 
-let jshintCmd = path.join(TEST_DIR, '/fixtures/node_modules/.bin/jshint');
+let eslintCmd = path.join(TEST_DIR, '/fixtures/node_modules/.bin/eslint');
 function testFile(command, _path) {
   _path = path.normalize(_path);
   return fs.accessAsync(_path, fs.R_OK).then(() => {
@@ -83,41 +75,19 @@ function testFile(command, _path) {
   });
 }
 
-function jshintDir(dir, name, folder) {
+function eslintDir(dir, name, folder) {
   if(!folder) folder = name;
   let endpointDir = path.join(dir, 'server/api', folder);
 
   let regFiles = fs.readdirAsync(endpointDir)
     .then(files => files.filter(file => minimatch(file, '**/!(*.spec|*.mock|*.integration).js', {dot: true})))
-    .map(file => testFile(jshintCmd, path.join('./server/api/', folder, file)));
+    .map(file => testFile(eslintCmd, path.join('./server/api/', folder, file)));
 
   let specFiles = fs.readdirAsync(endpointDir)
     .then(files => files.filter(file => minimatch(file, '**/+(*.spec|*.mock|*.integration).js', {dot: true})))
-    .map(file => testFile(`${jshintCmd} --config server/.jshintrc-spec`, path.join('./server/api/', folder, file)));
+    .map(file => testFile(`${eslintCmd} --env node,es6,mocha --global sinon,expect`, path.join('./server/api/', folder, file)));
 
   return Promise.all([regFiles, specFiles]);
-}
-function jscsDir(dir, name, folder) {
-  if(!folder) folder = name;
-  let endpointDir = path.join(dir, 'server/api', folder);
-
-  return fs.readdirAsync(endpointDir).then(files => {
-    return Promise.map(files, file => {
-      return fs.readFileAsync(path.join('server/api', folder, file), 'utf8').then(data => {
-        let results = jscs.checkString(data)
-        let errors = results.getErrorList();
-        if(errors.length === 0) {
-          return Promise.resolve();
-        } else {
-          errors.forEach(error => {
-            var colorizeOutput = true;
-            console.log(results.explainError(error, colorizeOutput) + '\n');
-          });
-          return Promise.reject();
-        }
-      });
-    });
-  });
 }
 
 var config;
@@ -128,10 +98,6 @@ describe('angular-fullstack:endpoint', function() {
     return Promise.all([
       runGen(defaultOptions).then(_dir => {
         genDir = _dir;
-
-        return fs.readFileAsync(path.join(genDir, '.jscsrc'), 'utf8').then(data => {
-          jscs.configure(JSON.parse(data));
-        });
       }),
       readJSON(path.join(TEST_DIR, 'fixtures/.yo-rc.json')).then(_config => {
         _config['generator-angular-fullstack'].insertRoutes = false;
@@ -150,9 +116,8 @@ describe('angular-fullstack:endpoint', function() {
         dir = _dir;
 
         return Promise.all([
-          copyAsync(path.join(genDir, '/server/.jshintrc'), './server/.jshintrc'),
-          copyAsync(path.join(genDir, '/server/.jshintrc-spec'), './server/.jshintrc-spec'),
-          copyAsync(path.join(genDir, '/.jscsrc'), './.jscsrc')
+          copyAsync(path.join(genDir, '/.eslintrc'), './.eslintrc'),
+          copyAsync(path.join(genDir, '/server/.eslintrc'), './server/.eslintrc')
         ]);
       });
     });
@@ -161,12 +126,8 @@ describe('angular-fullstack:endpoint', function() {
       assert.file(getExpectedFiles.endpoint('foo'));
     });
 
-    it('should pass jscs', function() {
-      return jscsDir(dir, 'foo').should.be.fulfilled();
-    });
-
     it('should pass lint', function() {
-      return jshintDir(dir, 'foo').should.be.fulfilled();
+      return eslintDir(dir, 'foo').should.be.fulfilled();
     });
   });
 
@@ -177,9 +138,8 @@ describe('angular-fullstack:endpoint', function() {
         dir = _dir;
 
         return Promise.all([
-          copyAsync(path.join(genDir, '/server/.jshintrc'), './server/.jshintrc'),
-          copyAsync(path.join(genDir, '/server/.jshintrc-spec'), './server/.jshintrc-spec'),
-          copyAsync(path.join(genDir, '/.jscsrc'), './.jscsrc')
+          copyAsync(path.join(genDir, '/.eslintrc'), './.eslintrc'),
+          copyAsync(path.join(genDir, '/server/.eslintrc'), './server/.eslintrc')
         ]);
       });
     });
@@ -188,12 +148,8 @@ describe('angular-fullstack:endpoint', function() {
       assert.file(getExpectedFiles.endpoint('Foo'));
     });
 
-    it('should pass jscs', function() {
-      return jscsDir(dir, 'Foo').should.be.fulfilled();
-    });
-
     it('should pass lint', function() {
-      return jshintDir(dir, 'Foo').should.be.fulfilled();
+      return eslintDir(dir, 'Foo').should.be.fulfilled();
     });
   });
 
@@ -204,9 +160,8 @@ describe('angular-fullstack:endpoint', function() {
         dir = _dir;
 
         return Promise.all([
-          copyAsync(path.join(genDir, '/server/.jshintrc'), './server/.jshintrc'),
-          copyAsync(path.join(genDir, '/server/.jshintrc-spec'), './server/.jshintrc-spec'),
-          copyAsync(path.join(genDir, '/.jscsrc'), './.jscsrc')
+          copyAsync(path.join(genDir, '/.eslintrc'), './.eslintrc'),
+          copyAsync(path.join(genDir, '/server/.eslintrc'), './server/.eslintrc')
         ]);
       });
     });
@@ -215,12 +170,8 @@ describe('angular-fullstack:endpoint', function() {
       assert.file(getExpectedFiles.endpoint('bar', 'foo/bar'));
     });
 
-    it('should pass jscs', function() {
-      return jscsDir(dir, 'foo', 'foo/bar').should.be.fulfilled();
-    });
-
     it('should pass lint', function() {
-      return jshintDir(dir, 'foo', 'foo/bar').should.be.fulfilled();
+      return eslintDir(dir, 'foo', 'foo/bar').should.be.fulfilled();
     });
   });
 
@@ -231,9 +182,8 @@ describe('angular-fullstack:endpoint', function() {
         dir = _dir;
 
         return Promise.all([
-          copyAsync(path.join(genDir, '/server/.jshintrc'), './server/.jshintrc'),
-          copyAsync(path.join(genDir, '/server/.jshintrc-spec'), './server/.jshintrc-spec'),
-          copyAsync(path.join(genDir, '/.jscsrc'), './.jscsrc')
+          copyAsync(path.join(genDir, '/.eslintrc'), './.eslintrc'),
+          copyAsync(path.join(genDir, '/server/.eslintrc'), './server/.eslintrc')
         ]);
       });
     });
@@ -242,12 +192,8 @@ describe('angular-fullstack:endpoint', function() {
       assert.file(getExpectedFiles.endpoint('foo-bar'));
     });
 
-    it('should pass jscs', function() {
-      return jscsDir(dir, 'foo-bar').should.be.fulfilled();
-    });
-
     it('should pass lint', function() {
-      return jshintDir(dir, 'foo-bar').should.be.fulfilled();
+      return eslintDir(dir, 'foo-bar').should.be.fulfilled();
     });
   });
 });
