@@ -1,13 +1,40 @@
 'use strict';
+// @flow
+<%_ if(filters.flow) { -%>
+type User = {
+  name: string;
+  email: string;
+  password: string;
+};
+<%_ } -%>
+<%_ if(filters.ts) { -%>
+interface User {
+  name: string;
+  email: string;
+  password: string;
+}
+<%_ } -%>
 
-class LoginController {
+export default class LoginController {
+  user: User = {
+    name: '',
+    email: '',
+    password: ''
+  };
+  errors = {login: undefined};
+  submitted = false;
+  Auth;
+  <%_ if(filters.ngroute) { -%>
+  $location;
+  <%_ } if(filters.uirouter) { -%>
+  $state;<% } %>
+
+  /*@ngInject*/
   constructor(Auth<% if (filters.ngroute) { %>, $location<% } %><% if (filters.uirouter) { %>, $state<% } %>) {
-    this.user = {};
-    this.errors = {};
-    this.submitted = false;
-
-    this.Auth = Auth;<% if (filters.ngroute) { %>
-    this.$location = $location;<% } if (filters.uirouter) { %>
+    this.Auth = Auth;
+    <%_ if(filters.ngroute) { -%>
+    this.$location = $location;
+    <%_ } if(filters.uirouter) { -%>
     this.$state = $state;<% } %>
   }
 
@@ -24,11 +51,8 @@ class LoginController {
         <% if (filters.ngroute) { %>this.$location.path('/');<% } %><% if (filters.uirouter) { %>this.$state.go('main');<% } %>
       })
       .catch(err => {
-        this.errors.other = err.message;
+        this.errors.login = err.message;
       });
     }
   }
 }
-
-angular.module('<%= scriptAppName %>')
-  .controller('LoginController', LoginController);
