@@ -1,9 +1,32 @@
 'use strict';
+// @flow
+<%_ if(filters.flow) { -%>
+type User = {
+  oldPassword: string;
+  newPassword: string;
+  confirmPassword: string;
+};
+<%_ } -%>
+<%_ if(filters.ts) { -%>
+interface User {
+  oldPassword: string;
+  newPassword: string;
+  confirmPassword: string;
+}
+<%_ } -%>
 
-class SettingsController {
-  errors = {};
+export default class SettingsController {
+  user: User = {
+    oldPassword: '',
+    newPassword: '',
+    confirmPassword: ''
+  };
+  errors = {other: undefined};
+  message = '';
   submitted = false;
+  Auth;
 
+  /*@ngInject*/
   constructor(Auth) {
     this.Auth = Auth;
   }
@@ -11,7 +34,7 @@ class SettingsController {
   changePassword(form) {
     this.submitted = true;
 
-    if (form.$valid) {
+    if(form.$valid) {
       this.Auth.changePassword(this.user.oldPassword, this.user.newPassword)
         .then(() => {
           this.message = 'Password successfully changed.';
@@ -24,6 +47,3 @@ class SettingsController {
     }
   }
 }
-
-angular.module('<%= scriptAppName %>')
-  .controller('SettingsController', SettingsController);
