@@ -1,13 +1,28 @@
 'use strict';
+// @flow
+class User {
+  _id: string = '';
+  name: string = '';
+  email: string = '';
+  role: string = '';
+  $promise = undefined;
+}
 
-(function() {
-
-function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
+export function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
+  'ngInject';
   var safeCb = Util.safeCb;
-  var currentUser = {};
+  var currentUser: User = new User();
   var userRoles = appConfig.userRoles || [];
+  /**
+   * Check if userRole is >= role
+   * @param {String} userRole - role of current user
+   * @param {String} role - role to check against
+   */
+  var hasRole = function(userRole, role) {
+    return userRoles.indexOf(userRole) >= userRoles.indexOf(role);
+  };
 
-  if ($cookies.get('token') && $location.path() !== '/logout') {
+  if($cookies.get('token') && $location.path() !== '/logout') {
     currentUser = User.get();
   }
 
@@ -17,14 +32,11 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
      * Authenticate user and save token
      *
      * @param  {Object}   user     - login info
-     * @param  {Function} callback - optional, function(error, user)
+     * @param  {Function} callback - function(error, user)
      * @return {Promise}
      */
-    login({email, password}, callback: Function) {
-      return $http.post('/auth/local', {
-        email: email,
-        password: password
-      })
+    login({email, password}, callback?: Function) {
+      return $http.post('/auth/local', { email, password })
         .then(res => {
           $cookies.put('token', res.data.token);
           currentUser = User.get();
@@ -46,17 +58,17 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
      */
     logout() {
       $cookies.remove('token');
-      currentUser = {};
+      currentUser = new User();
     },
 
     /**
      * Create a new user
      *
      * @param  {Object}   user     - user info
-     * @param  {Function} callback - optional, function(error, user)
+     * @param  {Function} callback - function(error, user)
      * @return {Promise}
      */
-    createUser(user, callback) {
+    createUser(user, callback?: Function) {
       return User.save(user,
         function(data) {
           $cookies.put('token', data.token);
@@ -74,14 +86,11 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
      *
      * @param  {String}   oldPassword
      * @param  {String}   newPassword
-     * @param  {Function} callback    - optional, function(error, user)
+     * @param  {Function} callback    - function(error, user)
      * @return {Promise}
      */
-    changePassword(oldPassword, newPassword, callback) {
-      return User.changePassword({ id: currentUser._id }, {
-        oldPassword: oldPassword,
-        newPassword: newPassword
-      }, function() {
+    changePassword(oldPassword, newPassword, callback?: Function) {
+      return User.changePassword({ id: currentUser._id }, { oldPassword, newPassword }, function() {
         return safeCb(callback)(null);
       }, function(err) {
         return safeCb(callback)(err);
@@ -90,18 +99,15 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
 
     /**
      * Gets all available info on a user
-     *   (synchronous|asynchronous)
      *
-     * @param  {Function|*} callback - optional, funciton(user)
-     * @return {Object|Promise}
+     * @param  {Function} [callback] - function(user)
+     * @return {Promise}
      */
-    getCurrentUser(callback) {
-      if (arguments.length === 0) {
-        return currentUser;
-      }
+    getCurrentUser(callback?: Function) {
+      var value = currentUser.hasOwnProperty('$promise')
+        ? currentUser.$promise
+        : currentUser;
 
-      var value = (currentUser.hasOwnProperty('$promise')) ?
-        currentUser.$promise : currentUser;
       return $q.when(value)
         .then(user => {
           safeCb(callback)(user);
@@ -113,18 +119,22 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
     },
 
     /**
-     * Check if a user is logged in
-     *   (synchronous|asynchronous)
+     * Gets all available info on a user
      *
-     * @param  {Function|*} callback - optional, function(is)
-     * @return {Bool|Promise}
+     * @return {Object}
      */
-    isLoggedIn(callback) {
-      if (arguments.length === 0) {
-        return currentUser.hasOwnProperty('role');
-      }
+    getCurrentUserSync() {
+      return currentUser;
+    },
 
-      return Auth.getCurrentUser(null)
+    /**
+     * Check if a user is logged in
+     *
+     * @param  {Function} [callback] - function(is)
+     * @return {Promise}
+     */
+    isLoggedIn(callback?: Function) {
+      return Auth.getCurrentUser(undefined)
         .then(user => {
           var is = user.hasOwnProperty('role');
           safeCb(callback)(is);
@@ -132,30 +142,42 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
         });
     },
 
+    /**
+     * Check if a user is logged in
+     *
+     * @return {Bool}
+     */
+    isLoggedInSync() {
+      return currentUser.hasOwnProperty('role');
+    },
+
      /**
       * Check if a user has a specified role or higher
-      *   (synchronous|asynchronous)
       *
       * @param  {String}     role     - the role to check against
-      * @param  {Function|*} callback - optional, function(has)
-      * @return {Bool|Promise}
+      * @param  {Function} [callback] - function(has)
+      * @return {Promise}
       */
-    hasRole(role, callback) {
-      var hasRole = function(r, h) {
-        return userRoles.indexOf(r) >= userRoles.indexOf(h);
-      };
-
-      if (arguments.length < 2) {
-        return hasRole(currentUser.role, role);
-      }
-
-      return Auth.getCurrentUser(null)
+    hasRole(role, callback?: Function) {
+      return Auth.getCurrentUser(undefined)
         .then(user => {
-          var has = (user.hasOwnProperty('role')) ?
-            hasRole(user.role, role) : false;
+          var has = user.hasOwnProperty('role')
+            ? hasRole(user.role, role)
+            : false;
+
           safeCb(callback)(has);
           return has;
         });
+    },
+
+    /**
+      * Check if a user has a specified role or higher
+      *
+      * @param  {String} role - the role to check against
+      * @return {Bool}
+      */
+    hasRoleSync(role) {
+      return hasRole(currentUser.role, role);
     },
 
      /**
@@ -170,6 +192,15 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
         .apply(Auth, [].concat.apply(['admin'], arguments));
     },
 
+     /**
+      * Check if a user is an admin
+      *
+      * @return {Bool}
+      */
+    isAdminSync() {
+      return Auth.hasRoleSync('admin');
+    },
+
     /**
      * Get auth token
      *
@@ -182,8 +213,3 @@ function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
 
   return Auth;
 }
-
-angular.module('<%= scriptAppName %>.auth')
-  .factory('Auth', AuthService);
-
-})();
