@@ -1,10 +1,12 @@
 'use strict';
+const $ = require('sprint-js');
+import OauthButtons from './index';
 
 describe('Directive: oauthButtons', function() {
 
   // load the directive's module and view
-  beforeEach(module('<%= scriptAppName %>'));
-  beforeEach(module('components/oauth-buttons/oauth-buttons.html'));
+  beforeEach(angular.mock.module(OauthButtons));
+  // beforeEach(angular.mock.module('components/oauth-buttons/oauth-buttons.html'));
 
   var element, parentScope, elementScope;
 
@@ -23,8 +25,8 @@ describe('Directive: oauthButtons', function() {
 
   it('should contain anchor buttons', function() {
     compileDirective('<oauth-buttons></oauth-buttons>');<% if (filters.jasmine) { %>
-    expect(element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>').length).toBeGreaterThan(0);<% } if (filters.mocha) { %>
-    <%= expect() %>element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>').length<%= to() %>.be.at.least(1);<% } %>
+    expect($(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>').length).toBeGreaterThan(0);<% } if (filters.mocha) { %>
+    <%= expect() %>$(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>').length<%= to() %>.be.at.least(1);<% } %>
   });
 
   it('should evaluate and bind the classes attribute to scope.classes', function() {
@@ -39,13 +41,13 @@ describe('Directive: oauthButtons', function() {
     // Add classes
     elementScope.classes = 'testClass1 testClass2';
     elementScope.$digest();<% if (filters.jasmine) { %>
-    expect(element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length).toBeGreaterThan(0);<% } if (filters.mocha) { %>
-    <%= expect() %>element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length<%= to() %>.be.at.least(1);<% } %>
+    expect($(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length).toBeGreaterThan(0);<% } if (filters.mocha) { %>
+    <%= expect() %>$(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length<%= to() %>.be.at.least(1);<% } %>
 
     // Remove classes
     elementScope.classes = '';
     elementScope.$digest();<% if (filters.jasmine) { %>
-    expect(element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length).toEqual(0);<% } if (filters.mocha) { %>
-    <%= expect() %>element.find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length<%= to() %>.equal(0);<% } %>
+    expect($(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length).toEqual(0);<% } if (filters.mocha) { %>
+    <%= expect() %>$(element[0]).find('a.btn<% if (filters.bootstrap) { %>.btn-social<% } %>.testClass1.testClass2').length<%= to() %>.equal(0);<% } %>
   });
 });
