@@ -1,8 +1,8 @@
-/* global io */
 'use strict';
 
-angular.module('<%= scriptAppName %>')
-  .factory('socket', function(socketFactory) {
+import io from 'socket.io-client';
+
+function Socket(socketFactory) {
     // socket.io now auto-configures its connection when we ommit a connection url
     var ioSocket = io('', {
       // Send auth token on connection, you will need to DI the Auth service above
@@ -68,4 +68,8 @@ angular.module('<%= scriptAppName %>')
         socket.removeAllListeners(modelName + ':remove');
       }
     };
-  });
+  }
+
+export default angular.module('<%= scriptAppName %>.socket', [])
+  .factory('socket', Socket)
+  .name;
