@@ -2,25 +2,25 @@ import passport from 'passport';
 import {Strategy as LocalStrategy} from 'passport-local';
 
 function localAuthenticate(User, email, password, done) {
-  <% if (filters.mongooseModels) { %>User.findOne({
+  <% if(filters.mongooseModels) { %>User.findOne({
     email: email.toLowerCase()
   }).exec()<% }
-     if (filters.sequelizeModels) { %>User.find({
+     if(filters.sequelizeModels) { %>User.find({
     where: {
       email: email.toLowerCase()
     }
   })<% } %>
     .then(user => {
-      if (!user) {
+      if(!user) {
         return done(null, false, {
           message: 'This email is not registered.'
         });
       }
       user.authenticate(password, function(authError, authenticated) {
-        if (authError) {
+        if(authError) {
           return done(authError);
         }
-        if (!authenticated) {
+        if(!authenticated) {
           return done(null, false, { message: 'This password is not correct.' });
         } else {
           return done(null, user);
@@ -30,11 +30,11 @@ function localAuthenticate(User, email, password, done) {
     .catch(err => done(err));
 }
 
-export function setup(User, config) {
+export function setup(User/*, config*/) {
   passport.use(new LocalStrategy({
     usernameField: 'email',
     passwordField: 'password' // this is the virtual field on the model
-  }, function(email, password, done) {<% if (filters.models) { %>
+  }, function(email, password, done) {<% if(filters.models) { %>
     return localAuthenticate(User, email, password, done);
 <% } %>  }));
 }
