@@ -13,7 +13,7 @@ module.exports[404] = function pageNotFound(req, res) {
 
   res.status(result.status);
   res.render(viewFilePath, {}, function(err, html) {
-    if (err) {
+    if(err) {
       return res.status(result.status).json(result);
     }
 
