@@ -1,30 +1,30 @@
 'use strict';
+/*eslint no-process-env:0*/
 
 // Production specific configuration
 // =================================
 module.exports = {
   // Server IP
-  ip:     process.env.OPENSHIFT_NODEJS_IP ||
-          process.env.IP ||
-          undefined,
+  ip: process.env.OPENSHIFT_NODEJS_IP
+    || process.env.ip
+    || undefined,
 
   // Server port
-  port:   process.env.OPENSHIFT_NODEJS_PORT ||
-          process.env.PORT ||
-          <%= prodPort %><% if (filters.mongoose) { %>,
+  port: process.env.OPENSHIFT_NODEJS_PORT
+    || process.env.port
+    || <%= prodPort %><% if(filters.mongoose) { %>,
 
   // MongoDB connection options
   mongo: {
-    uri:  process.env.MONGODB_URI ||
-          process.env.MONGOHQ_URL ||
-          process.env.OPENSHIFT_MONGODB_DB_URL +
-          process.env.OPENSHIFT_APP_NAME ||
-          'mongodb://localhost/<%= lodash.slugify(appname) %>'
+    uri: process.env.MONGODB_URI
+      || process.env.MONGOHQ_URL
+      || process.env.OPENSHIFT_MONGODB_DB_URL + process.env.OPENSHIFT_APP_NAME
+      || 'mongodb://localhost/<%= lodash.slugify(appname) %>'
   }<% } if (filters.sequelize) { %>,
 
   sequelize: {
-    uri:  process.env.SEQUELIZE_URI ||
-          'sqlite://',
+    uri: process.env.SEQUELIZE_URI
+      || 'sqlite://',
     options: {
       logging: false,
       storage: 'dist.sqlite',
