@@ -11,9 +11,9 @@ var events = ['save', 'remove'];
 
 export function register(socket) {
   // Bind model events to socket events
-  for (var i = 0, eventsLength = events.length; i < eventsLength; i++) {
+  for(var i = 0, eventsLength = events.length; i < eventsLength; i++) {
     var event = events[i];
-    var listener = createListener('<%= cameledName %>:' + event, socket);
+    var listener = createListener(`<%= cameledName %>:${event}`, socket);
 
     <%= classedName %>Events.on(event, listener);
     socket.on('disconnect', removeListener(event, listener));
